Update core form handling servlet to validate captcha
Updates the `CoreFormHanldingServlet` to validate that a captcha
response is present and valid if required, prior to processing the
form POST request.

If the captcha is required but not present, or invalid, then a
`forbidden` status code is returned.

Update tests for `CoreFormHandlingServlet`

Updates tests to verify the proper handling of the request
by the servlet when the captcha is:
 1) not required
 2) required by component property; and,
   a) is valid
   b) is not valid
 3) required by policy; and,
   a) is valid
   b) is not valid

Add JavaDocs to undocumented members of `CoreFormHandlingServlet`

Adds JavaDoc comments to all undocumented internal members of the
`CoreFormHandlingServlet` class.