Add utility method for getting child components and component exporters
Adds general utility methods to the `ComponentUtils` class for the purpose
of finding child resources that are components, and child resources that
are component exporters.

This logic is moved to the utility class because it is likely to be reused
by numerous container components, and is likely to be useful for custom
client container models.