Convert `AbstractContainerImpl` to use `ComponentUtils`
Updates the `Container` model implementations to use the newly
added helper methods`ComponentUtils#getComponentModels()` and
`ComponentUtils#getChildComponents()`.

The return type of `AbstractContainerImpl#getExportedItems()`
is changed to `LinkedHashMap` to indicate the ordering of these
items is critically important.

The `AbstractContainerImpl#getExportedItems()` method is made
abstract, instead of relying on other abstract methods to
provide the data. This clarifies the application flow without
introducing undue duplication of logic.