Allow button icons to be restricted to pre-determined options
Updates the Button V2 design dialog to allow the author to restrict
the "icon" options to a pre-determined list, which can be configured
in the component policy.

Updates the Button V2 edit dialog to present a select field instead of
a text field when the policy restricts the allowed icons.